Improve Native time on Daily action
++ Run only http/http-minimum in Windows native only
++ Disable AnnotationScheduledJobsMySqlQuartzIT on Native as it takes around 30 min to finish
++ There were some duplicated test cases in QuarkusCliVersionIT and QuarkusCliCreateJvmApplicationIT
